// frontend/src/app/core/api/model/changePasswordRequest.ts
/**
 * Le Tacos Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ChangePasswordRequestDto { 
    currentPassword: string;
    newPassword: string;
}